fix up about us page
